Fixes AsyncLock invalid state (#5)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

The AsyncLock was entering an invalid state when a locked task was
canceled. This occurred because the unlock operation failed to skip the
non-pending tasks, resulting in an error of invalid state that persisted
indefinitely.

Co-authored-by: brennobemoura <37243584+brennobemoura@users.noreply.github.com>